+++ src/OpenTelemetry.Exporter.Zipkin/CHANGELOG.md
@@ -5,6 +5,13 @@
 * Zipkin will now set the `error` tag when `otel.status_code` is set to `Error`.
   ([#1579](https://github.com/open-telemetry/opentelemetry-dotnet/pull/1579))
 
+* Zipkin will no longer send the `otel.status_code` tag if the value is `Unset`.
+  ([#1609](https://github.com/open-telemetry/opentelemetry-dotnet/pull/1609))
+
+* Zipkin bool tag values will now be sent as `true`/`false` instead of
+  `True`/`False`.
+  ([#1609](https://github.com/open-telemetry/opentelemetry-dotnet/pull/1609))
+
 ## 1.0.0-rc1.1
 
 Released 2020-Nov-17